sections/04_react_agents_from_scratch/projects/react-langchain/pyproject.toml
[project]
name = "react-agent-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.11.0",
    "langchain>=0.2.0,<0.3",
    "langchain-google-genai>=1.0.0,<2.0",
    "langsmith>=0.1.0,<0.2.0",
    "python-dotenv>=1.0.0",
]
